chore: Move the async big-segment status provider out of _common
AsyncBigSegmentStoreStatusProviderImpl is async-only (used only by
async_big_segments), so it belongs there, not in big_segments_common
(which is for logic shared by the sync and async managers).
